Fix cookie contest pseudocode
The cookie contest pseudocode can sometimes produce RESPONDER -
RESPONDER type of peers, because bit 31 is used both as a weight in a
64-bit 2's complement integer, and a sign bit. See the discussion here
for a detailed brief:

https://github.com/Haivision/srt/issues/3208

This should be fixed asap in the RFC such that implementors do not end
up with this bug in their protocol stack.

We propose to avoid the use of bitmasks to
 1) make the intention clearer
 2) let the compiler handle the low-level details

We also propose to add a statement, reminding implementors of the
importance of considering endianness of the host and the network before
comparing the cookies.

squash